fix: Address multiple audit findings in ProtocolManager   Security fixes:   - Enforce ips and timestampStart immutability in setGuardedPriceConfig     to prevent price manipulation via guard parameter changes   - Add TooEarlyInPeriod check requiring updateManagementConfig calls     to occur in final 1/3 of period, preventing mid-period limit changes     that could trap managers in inconsistent adjustment states   - Move limit validation inside hasAuthority block since checks are     irrelevant when removing authority (limits set to zero)
  Overflow prevention (struct packing preserved at 5 slots):
  - Upgrade PeriodAdjustments.decayPerAdjustment from int8 to int16
  - Upgrade PeriodAdjustments.vertexMultiplierMax from int16 to int24
  - Upgrade price guard fields from int88/uint88 to int96/uint96
  - Update corresponding PeriodLimits types to match

  Documentation:
  - Fix inaccurate comments on MAXIMUM_DECAY_RATE_LIMIT and
    MAXIMUM_VERTEX_MULTIPLIER_MAX_LIMIT constants
  - Add comment explaining rounding behavior in _perSecondToBPS

  Tests:
  - Add _warpToValidManagementConfigWindow helper for time restrictions
  - Add test_updateManagementConfig_fail_tooEarlyInPeriod test case
  - Update expected errors for ips/timestampStart validation
  - Update type references throughout test suite

+++ tests/architecture/ProtocolManager/function/UpdateDynamicIRM.t.sol
@@ -591,8 +591,8 @@ contract TestProtocolManagerUpdateDynamicIRM is TestProtocolManagerBase {
             int64 vertexInterestRateAdj,
             int64 vertexStartAdj,
             int16 adjustmentVelocityAdj,
-            int8 decayPerAdjustmentAdj,
-            int16 vertexMultiplierMaxAdj
+            int16 decayPerAdjustmentAdj,
+            int24 vertexMultiplierMaxAdj
         ) = protocolManager.getMarketPeriodAdjustments(address(dynamicIRM), periodTimestamp);
 
         // All adjustments should be exact
@@ -635,8 +635,8 @@ contract TestProtocolManagerUpdateDynamicIRM is TestProtocolManagerBase {
             int64 vertexInterestRateAdj,
             int64 vertexStartAdj,
             int16 adjustmentVelocityAdj,
-            int8 decayPerAdjustmentAdj,
-            int16 vertexMultiplierMaxAdj
+            int16 decayPerAdjustmentAdj,
+            int24 vertexMultiplierMaxAdj
         ) = protocolManager.getMarketPeriodAdjustments(address(dynamicIRM), periodTimestamp);
 
         // All adjustments should be exact
@@ -690,8 +690,8 @@ contract TestProtocolManagerUpdateDynamicIRM is TestProtocolManagerBase {
             int64 vertexInterestRateAdj,
             int64 vertexStartAdj,
             int16 adjustmentVelocityAdj,
-            int8 decayPerAdjustmentAdj,
-            int16 vertexMultiplierMaxAdj
+            int16 decayPerAdjustmentAdj,
+            int24 vertexMultiplierMaxAdj
         ) = protocolManager.getMarketPeriodAdjustments(address(dynamicIRM), periodTimestamp);
 
         assertEq(baseInterestRateAdj, 200, "Base rate at limit");
@@ -769,8 +769,8 @@ contract TestProtocolManagerUpdateDynamicIRM is TestProtocolManagerBase {
             int64 vertexInterestRateAdj,
             int64 vertexStartAdj,
             int16 adjustmentVelocityAdj,
-            int8 decayPerAdjustmentAdj,
-            int16 vertexMultiplierMaxAdj
+            int16 decayPerAdjustmentAdj,
+            int24 vertexMultiplierMaxAdj
         ) = protocolManager.getMarketPeriodAdjustments(address(dynamicIRM), periodTimestamp);
 
         // All adjustments should be exactly zero
